chore: fix checkstyle and restructure.
