--- spring-cloud-gcp-autoconfigure/src/main/java/com/google/cloud/spring/autoconfigure/trace/TraceRuntimeHints.java
@@ -1,44 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-
-package com.google.cloud.spring.autoconfigure.trace;
-
-import com.google.protobuf.Value;
-import io.micrometer.observation.aop.ObservedAspect;
-import java.lang.reflect.Method;
-import java.util.Arrays;
-import org.aspectj.lang.ProceedingJoinPoint;
-import org.springframework.aot.hint.ExecutableMode;
-import org.springframework.aot.hint.MemberCategory;
-import org.springframework.aot.hint.RuntimeHints;
-import org.springframework.aot.hint.RuntimeHintsRegistrar;
-import org.springframework.aot.hint.TypeReference;
-import org.springframework.util.ReflectionUtils;
-
-public class TraceRuntimeHints implements RuntimeHintsRegistrar {
-
-  @Override
-  public void registerHints(RuntimeHints hints, ClassLoader classLoader) {
-    Method method = ReflectionUtils.findMethod(ObservedAspect.class, "observeMethod", ProceedingJoinPoint.class);
-    hints.reflection().registerMethod(method, ExecutableMode.INVOKE);
-
-    // hints.reflection().registerTypes(
-    //     Arrays.asList(TypeReference.of(Value.class), TypeReference.of(Value.Builder.class)),
-    //     hint -> hint.withMembers(MemberCategory.DECLARED_FIELDS,
-    //         MemberCategory.INVOKE_DECLARED_METHODS));
-  }
-}
