+++ spring-cloud-gcp-autoconfigure/src/main/resources/META-INF/spring/aot.factories
@@ -1,2 +1,2 @@
 org.springframework.aot.hint.RuntimeHintsRegistrar=\
-com.google.cloud.spring.autoconfigure.trace.TraceRuntimeHints
+com.google.cloud.spring.autoconfigure.trace.aot.TraceRuntimeHints
