+++ spring-cloud-gcp-samples/spring-cloud-gcp-trace-sample/src/test/java/com/example/TestRuntimeHints.java
@@ -28,7 +28,6 @@ public class TestRuntimeHints implements RuntimeHintsRegistrar {
 
   @Override
   public void registerHints(RuntimeHints hints, ClassLoader classLoader) {
-
     hints
         .reflection()
         .registerTypes(
